# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: OPES School Software API
  description: >-
    Read-only integration surface over the modular monolith. REST is a thin
    adapter over the same module internals as the Livewire screens
    (00-core 6.1); it enforces the same permission gates and never exposes a
    write path in v1.
  version: 1.0.0
servers:
  - url: /api
security:
  - bearerToken: []
paths:
...
  /v1/me/profile:
    patch:
      operationId: updateOwnProfile
      summary: Edit own contact details (row 29)
      description: >-
        Row 29 is the only place a guardian writes about themselves, and row 30
        is what makes it safe: an attempt to set `has_custody`,
        `receives_reports`, `receives_invoices`, `is_fee_payer`, `status`,
        `is_archived` or `portal_user_id` is a SECURITY event - audited, then
        403. A guardian who could edit their own authorization flags could grant
        themselves every other row in the matrix. The permitted set is a closed
        allow-list, never a deny-list, so a column added to `guardians` next
        year is not silently writable. Requires `portal.write`.
      responses:
        '200': { description: The updated contact details. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /v1/me/devices/push:
    post:
      operationId: registerPushDevice
      summary: Register this device for push
      description: >-
        Web Push and Expo push are the same registration with different key
        material, discriminated by `platform` (spec 7). Upserts on `endpoint`,
        so an app re-subscribing after a permission refresh does not pile up
        duplicate rows that then receive duplicate pushes. Requires
        `portal.write`.
      responses:
        '201': { description: Registered. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    delete:
      operationId: unregisterPushDevice
      summary: Unregister this device
      description: >-
        Idempotent on purpose - unregistering a device that was already gone is
        a success, because an app signing out must not be able to fail here.
        Scoped by user id inside the Action: `endpoint` is unique table-wide, so
        an unscoped delete would let any signed-in guardian unsubscribe a
        stranger's device by quoting its endpoint.
      responses:
        '200': { description: Unregistered. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: Sanctum personal access token.
...
  responses:
    Unauthenticated:
      description: Missing or invalid token.
    Forbidden:
      description: >-
        The user lacks the required permission, or the token lacks the
        required ability.
...
    ValidationFailed:
      description: >-
        The request body failed validation. On the sign-in endpoint this is
        also the single, deliberately uninformative answer to every credential
        failure.
    TooManyRequests:
      description: >-
        A rate limit is exhausted - the 60/min `api` budget, the 5/min
        `auth-mobile` sign-in budget, or the 120/min `api-portal` device budget.
...